Refactor reusable packages in src and update imports
