+++ app/models/__init__.py
@@ -1,0 +1,3 @@
+from src.models import WalletData
+
+__all__ = ["WalletData"]
