+++ app/utils/db.py
@@ -1,21 +1,6 @@
 
-import os
-from functools import lru_cache
+"""Wrapper around ``src.utils`` database helpers."""
 
-try:
-    from motor.motor_asyncio import AsyncIOMotorClient
-except Exception:  # pragma: no cover - optional dependency
-    AsyncIOMotorClient = None  # type: ignore
+from src.utils import get_db
 
-MONGODB_URI = os.getenv("MONGODB_URI", "mongodb://localhost:27017")
-
-
-@lru_cache()
-def get_client() -> AsyncIOMotorClient:
-    if AsyncIOMotorClient is None:
-        raise ImportError("motor is required for database operations")
-    return AsyncIOMotorClient(MONGODB_URI)
-
-
-def get_db():
-    return get_client().foundlab
+__all__ = ["get_db"]
