+++ src/__init__.py
@@ -1,0 +1,1 @@
+"""Shared reusable logic for FoundLab services."""
