+++ src/main.py
@@ -1,6 +1,5 @@
-
 from fastapi import FastAPI
-from app.routers import score
+from src.routers import score
 
 app = FastAPI()
 
@@ -8,5 +7,5 @@ app.include_router(score.router)
 
 
 @app.get("/")
-def root():
+async def root():
     return {"status": "FoundLab DEP API online"}
