+++ src/models/__init__.py
@@ -1,0 +1,3 @@
+from .schemas import WalletData
+
+__all__ = ["WalletData"]
