+++ src/routers/__init__.py
@@ -1,0 +1,3 @@
+from .score import router
+
+__all__ = ["router"]
