+++ src/routers/score.py
@@ -1,0 +1,11 @@
+from fastapi import APIRouter
+from src.models import WalletData
+from src.services import calculate_score as engine_calculate_score
+
+router = APIRouter(prefix="/score")
+
+
+@router.post("/")
+async def calculate_score(data: WalletData):
+    """Return a reputation score using ``src`` utilities."""
+    return engine_calculate_score(data)
