+++ src/services/__init__.py
@@ -1,0 +1,9 @@
+from .engine import calculate_score, bayes_px, compute_probabilities
+from .score_engine import calculate as calculate_from_flags
+
+__all__ = [
+    "calculate_score",
+    "bayes_px",
+    "compute_probabilities",
+    "calculate_from_flags",
+]
