+++ src/utils/__init__.py
@@ -1,0 +1,3 @@
+from .db import get_db
+
+__all__ = ["get_db"]
